refactor: englify Langue state to Languages
